training/heterogeneous-clusters/pt.grpc.local/requirements.txt
torch
torchvision
grpcio-tools
tensorboard
torch-tb-profiler